Allow usage without loading minitest/spec
Currently it's not possible for minitest/test users to require
minitest-hooks without loading minitest/spec, and minitest/spec adds a
lot of monkey patches, which minitest/test users might want to avoid.
This change allows minitest/test users to require "minitest/hooks/test",
which doesn't load minitest/spec.
